add options to doctors screen
